Remove no longer needed font awesome related code.
