+++ graylog2-web-interface/src/components/common/Icon/Icon.md
