+++ graylog2-web-interface/src/components/common/Icon/types.ts
@@ -15,15 +15,7 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 
-import React from 'react';
-import find from 'lodash/find';
-import type { RotateProp } from '@fortawesome/fontawesome-svg-core';
-import styled, { css } from 'styled-components';
-
-import deprecationNotice from 'util/deprecationNotice';
-import loadAsync from 'routing/loadAsync';
-
-import compareIconNames from './icon-fallback';
+export type SizeProp = 'xs' | 'sm' | 'lg' | 'xl' | '2x' | '3x' | '4x' | '5x'
 
 export type IconName = 'tamper_detection_off' |
   'crisis_alert' |
@@ -3235,162 +3227,3 @@ export type IconName = 'tamper_detection_off' |
   'keyboard_capslock' |
   'pin_invoke' |
   'cleaning';
-
-export type SizeProp = 'xs' | 'sm' | 'lg' | 'xl' | '2x' | '3x' | '4x' | '5x'
-
-const CustomFontAwesomeIcon = loadAsync(() => import('./CustomFontAwesomeIcon'));
-
-type IconTypes = 'brand' | 'regular' | 'solid';
-
-const removeFaPrefix = (name: string) => name.replace(/^fa-/, ''); // remove "fa-" prefix if it exists
-
-const updateLegacyName = (icon: string) => {
-  const v4icon = find(compareIconNames, { v4: icon });
-  const iconName = (v4icon && v4icon.v5) || icon;
-
-  if (v4icon) {
-    deprecationNotice(`You have used a deprecated \`Icon\` name. \`${icon}\` should be \`${iconName}\``);
-  }
-
-  return iconName;
-};
-
-const cleanIconName = (name: string, type: IconTypes) => {
-  const iconName = removeFaPrefix(name);
-
-  if (type !== 'brand') {
-    return updateLegacyName(iconName);
-  }
-
-  return iconName;
-};
-
-const getPrefixForType = (type: IconTypes) => {
-  switch (type) {
-    case 'brand':
-      return 'fab';
-    case 'regular':
-      return 'far';
-    case 'solid':
-    default:
-      return 'fas';
-  }
-};
-
-const sizeMap = {
-  xs: '.75em',
-  sm: '.875em',
-  lg: '1.25em',
-  xl: '1.5em',
-  '2x': '2em',
-  '3x': '3em',
-  '4x': '4em',
-  '5x': '5em',
-};
-
-const StyledSpan = styled.span<{ $size: string }>(({ theme, $size }) => css`
-  font-variation-settings: 'opsz' 48, 'wght' 700;
-  font-size: ${sizeMap[$size] ?? 'inherit'};
-`);
-
-type Props = {
-  className?: string,
-  'data-testid'?: string,
-  /** Name of Font Awesome 5 Icon without `fa-` prefix */
-  name: IconName,
-  rotation?: RotateProp,
-  size?: SizeProp,
-  spin?: boolean,
-  /**
-   * Name of icon type, the brand type is needed for all brand icons.
-   * The type regular is needed to outlined icon.
-   * Not all icons can be outlined.
-   * */
-  type?: IconTypes,
-  fixedWidth?: boolean,
-  inverse?: boolean,
-  style?: React.CSSProperties,
-  onClick?: (event: React.MouseEvent<SVGSVGElement>) => void,
-  onMouseEnter?: (event: React.MouseEvent<SVGSVGElement>) => void,
-  onMouseLeave?: (event: React.MouseEvent<SVGSVGElement>) => void,
-  onFocus?: (event: React.FocusEvent<SVGSVGElement>) => void,
-  tabIndex?: number,
-  title?: string,
-}
-
-/**
- * Component that renders an icon or glyph.
- * Uses Font Awesome 5 : https://fontawesome.com/icons
- *
- * No need to pass `fa` or `fa-` prefixes, just the name of the icon
- * Visit [React FontAwesome Features](https://github.com/FortAwesome/react-fontawesome#features) for more information.
- */
-
-const Icon = ({
-  name,
-  type,
-  size,
-  className,
-  rotation,
-  spin,
-  fixedWidth,
-  inverse,
-  style,
-  'data-testid': testId,
-  onClick,
-  onMouseEnter,
-  onMouseLeave,
-  onFocus,
-  tabIndex,
-  title,
-}: Props) => {
-  const iconName = cleanIconName(name, type);
-  const prefix = getPrefixForType(type);
-
-  return (
-    <StyledSpan className={`material-symbols-outlined ${className}`}
-                data-testid={testId}
-                $size={size}
-                style={style}>
-      {iconName}
-    </StyledSpan>
-  );
-
-  return (
-    <CustomFontAwesomeIcon className={className}
-                           data-testid={testId}
-                           fixedWidth={fixedWidth}
-                           icon={{ prefix, iconName }}
-                           inverse={inverse}
-                           rotation={rotation}
-                           size={size}
-                           spin={spin}
-                           style={style}
-                           onClick={onClick}
-                           onMouseEnter={onMouseEnter}
-                           onMouseLeave={onMouseLeave}
-                           tabIndex={tabIndex}
-                           title={title}
-                           onFocus={onFocus} />
-  );
-};
-
-Icon.defaultProps = {
-  className: undefined,
-  'data-testid': undefined,
-  fixedWidth: false,
-  inverse: false,
-  rotation: undefined,
-  size: undefined,
-  spin: false,
-  style: undefined,
-  type: 'solid',
-  onClick: undefined,
-  onMouseEnter: undefined,
-  onMouseLeave: undefined,
-  onFocus: undefined,
-  tabIndex: undefined,
-  title: undefined,
-};
-
-export default Icon;
